Hide modules whose distributions we're exposing directly throguh rand::distributions. Also move a couple of distribution helper functions into utils.rs.
